Fix Google data retrieval half hour bug
Retrieving Google data with an interval of more than 30 min becomes an
issue because the market opens at 9:30 and closes at 4. This means that
depending on the current time (time floor is on the hour or on the half
hour) the first time of the day will be 9:30 or 10. Data retrieval
should have a minimum interval of 30 min to get all data and filtering
should occur afterwards.