Better handle Algorithm Request timeframes not ending in now
When a result set from Alfred is recieved Engine fetches actuals for
that time period. Requests where the whole time period is in the past
get actuals for the next cycle after that period, but that's not what
we're predicting on. Instead, we just need the value for that date from
Alfred as it's actually predicting the next cycle in the future.